+++ src/helpers/ChartHelper.php
@@ -56,49 +56,71 @@ class ChartHelper
 			$intervalUnit = self::getRunChartIntervalUnit($startDate, $endDate);
 		}
 
+
+        // Convert timezone for SQL request
+
+        $utc = new DateTime();
+        $timezoneId = craft()->getTimeZone();
+        $timezone = new \DateTimeZone($timezoneId);
+
+        $offsetInSeconds = $timezone->getOffset($utc);
+        $offsetInHours = round($offsetInSeconds / 60 / 60);
+
+        $fromTz = '+00:00';
+        $toTz = ($offsetInHours >= 0 ? '+' : '').$offsetInHours.':00';
+
+        $tzDateColumn = "CONVERT_TZ({$dateColumn},'{$fromTz}','{$toTz}')";
+
+
+        // Prepare the query
+
 		switch ($intervalUnit)
 		{
 			case 'year':
 			{
 				$sqlDateFormat = '%Y-01-01';
 				$phpDateFormat = 'Y-01-01';
-				$sqlGroup = "YEAR({$dateColumn})";
+				$sqlGroup = "YEAR({$tzDateColumn})";
 				break;
 			}
 			case 'month':
 			{
 				$sqlDateFormat = '%Y-%m-01';
 				$phpDateFormat = 'Y-m-01';
-				$sqlGroup = "YEAR({$dateColumn}), MONTH({$dateColumn})";
+				$sqlGroup = "YEAR({$tzDateColumn}), MONTH({$tzDateColumn})";
 				break;
 			}
 			case 'day':
 			{
 				$sqlDateFormat = '%Y-%m-%d';
 				$phpDateFormat = 'Y-m-d';
-				$sqlGroup = "YEAR({$dateColumn}), MONTH({$dateColumn}), DAY({$dateColumn})";
+				$sqlGroup = "YEAR({$tzDateColumn}), MONTH({$tzDateColumn}), DAY({$tzDateColumn})";
 				break;
 			}
 			case 'hour':
 			{
 				$sqlDateFormat = '%Y-%m-%d %H:00:00';
 				$phpDateFormat = 'Y-m-d H:00:00';
-				$sqlGroup = "YEAR({$dateColumn}), MONTH({$dateColumn}), DAY({$dateColumn}), HOUR({$dateColumn})";
+				$sqlGroup = "YEAR({$tzDateColumn}), MONTH({$tzDateColumn}), DAY({$tzDateColumn}), HOUR({$tzDateColumn})";
 				break;
 			}
 		}
 
+
 		// Execute the query
+
 		$results = $query
-			->addSelect("DATE_FORMAT({$dateColumn}, '{$sqlDateFormat}') as date")
+            ->addSelect("DATE_FORMAT({$tzDateColumn}, '{$sqlDateFormat}') as date")
 			->andWhere(
-				array('and', $dateColumn.' >= :startDate', $dateColumn.' < :endDate'),
-				array(':startDate' => $startDate->format(DateTime::MYSQL_DATETIME, DateTime::UTC), ':endDate' => $endDate->format(DateTime::MYSQL_DATETIME, DateTime::UTC)))
+				array('and', "{$tzDateColumn} >= :startDate", "{$tzDateColumn} < :endDate"),
+				array(':startDate' => $startDate->format(DateTime::MYSQL_DATETIME), ':endDate' => $endDate->format(DateTime::MYSQL_DATETIME)))
 			->group($sqlGroup)
-			->order($dateColumn.' asc')
+			->order("{$tzDateColumn} asc")
 			->queryAll();
 
-		// Assembe the data
+
+		// Assemble the data
+
 		$rows = array();
 
 		$cursorDate = $startDate;
@@ -107,10 +129,9 @@ class ChartHelper
 		while ($cursorDate->getTimestamp() < $endTimestamp)
 		{
 			// Do we have a record for this date?
-			$cursorDateUtc = $cursorDate->format($phpDateFormat, DateTime::UTC);
 			$formattedCursorDate = $cursorDate->format($phpDateFormat);
 
-			if (isset($results[0]) && $results[0]['date'] == $cursorDateUtc)
+			if (isset($results[0]) && $results[0]['date'] == $formattedCursorDate)
 			{
 				$value = (float) $results[0]['value'];
 				array_shift($results);